+++ simple_history/models.py
@@ -1,28 +1,24 @@
 from __future__ import unicode_literals
 
-import threading
 import copy
-import warnings
-import django
+import importlib
+import threading
+
 from django.db import models, router
-from django.db.models import loading, Q
+from django.db.models import Q
 from django.db.models.fields.proxy import OrderWrt
-from django.db.models.fields.related import RelatedField
 from django.conf import settings
 from django.contrib import admin
-from django.utils import importlib, six
+from django.utils import six
 from django.utils.encoding import python_2_unicode_compatible
 from django.utils.encoding import smart_text
 from django.utils.timezone import now
 from django.utils.translation import string_concat
 
-from simple_history import register
-from .manager import HistoryDescriptor
-
 try:
     from django.apps import apps
 except ImportError:  # Django < 1.7
-    apps = None
+    from django.db.models import get_app
 try:
     from south.modelsinspector import add_introspection_rules
 except ImportError:  # south not present
@@ -30,6 +26,10 @@ except ImportError:  # south not present
 else:  # south configuration for CustomForeignKeyField
     add_introspection_rules(
         [], ["^simple_history.models.CustomForeignKeyField"])
+
+from . import exceptions
+from simple_history import register
+from .manager import HistoryDescriptor
 
 ALL_M2M_FIELDS = object()
 
@@ -45,9 +45,12 @@ def not_registered(model):
 class HistoricalRecords(object):
     thread = threading.local()
 
-    def __init__(self, verbose_name=None, bases=(models.Model,), user_related_name='+', m2m_fields=None):
+    def __init__(self, verbose_name=None, bases=(models.Model,),
+                 user_related_name='+', table_name=None, inherit=False, m2m_fields=None):
         self.user_set_verbose_name = verbose_name
         self.user_related_name = user_related_name
+        self.table_name = table_name
+        self.inherit = inherit
         self.m2m_fields = m2m_fields
         try:
             if isinstance(bases, six.string_types):
@@ -112,9 +115,13 @@ class HistoricalRecords(object):
             pass
         else:
             if hint_class is not sender:  # set in concrete
-                if not (hint_class._meta.abstract
-                        and issubclass(sender, hint_class)):  # set in abstract
+                if not (self.inherit and issubclass(sender, hint_class)):  # set in abstract
                     return
+        if hasattr(sender._meta, 'simple_history_manager_attribute'):
+            raise exceptions.MultipleRegistrationsError('{}.{} registered multiple times for history tracking.'.format(
+                sender._meta.app_label,
+                sender._meta.object_name,
+            ))
         if sender._meta.proxy:
             original_class = [base_class for base_class in sender.__bases__ if base_class._meta.abstract is False][0]
             # Parent model must be registered before the proxy model is
@@ -160,20 +167,22 @@ class HistoricalRecords(object):
             # registered under different app
             attrs['__module__'] = self.module
         elif app_module != self.module:
-            if apps is None:  # Django < 1.7
-                # has meta options with app_label
-                app = loading.get_app(model._meta.app_label)
-                attrs['__module__'] = app.__name__  # full dotted name
-            else:
+            try:
                 # Abuse an internal API because the app registry is loading.
                 app = apps.app_configs[model._meta.app_label]
-                attrs['__module__'] = app.name      # full dotted name
+            except NameError:  # Django < 1.7
+                models_module = get_app(model._meta.app_label).__name__
+            else:
+                models_module = app.name
+            attrs['__module__'] = models_module
 
         fields = self.copy_fields(model)
         attrs.update(fields)
         attrs.update(self.get_extra_fields(model, fields))
         # type in python2 wants str as a first argument
         attrs.update(Meta=type(str('Meta'), (), self.get_meta_options(model)))
+        if self.table_name is not None:
+            attrs['Meta'].db_table = self.table_name
         name = 'Historical%s' % model._meta.object_name
         registered_models[model._meta.db_table] = model
         return python_2_unicode_compatible(
@@ -187,20 +196,25 @@ class HistoricalRecords(object):
         fields = {}
         for field in model._meta.fields:
             field = copy.copy(field)
-            field.rel = copy.copy(field.rel)
+            try:
+                field.remote_field = copy.copy(field.remote_field)
+            except AttributeError:
+                field.rel = copy.copy(field.rel)
             if isinstance(field, OrderWrt):
                 # OrderWrt is a proxy field, switch to a plain IntegerField
                 field.__class__ = models.IntegerField
             if isinstance(field, models.ForeignKey):
                 old_field = field
-                field_arguments = {}
+                field_arguments = {'db_constraint': False}
                 if (getattr(old_field, 'one_to_one', False) or
                         isinstance(old_field, models.OneToOneField)):
                     FieldType = models.ForeignKey
                 else:
                     FieldType = type(old_field)
-                if django.get_version() >= "1.6":
-                    field_arguments['db_constraint'] = False
+                if getattr(old_field, 'to_fields', []):
+                    field_arguments['to_field'] = old_field.to_fields[0]
+                if getattr(old_field, 'db_column', None):
+                    field_arguments['db_column'] = old_field.db_column
                 field = FieldType(
                     old_field.rel.to,
                     related_name='+',
@@ -228,17 +242,16 @@ class HistoricalRecords(object):
         def revert_url(self):
             """URL for this change in the default admin site."""
             opts = model._meta
-            try:
-                app_label, model_name = opts.app_label, opts.model_name
-            except AttributeError:  # Django < 1.7
-                app_label, model_name = opts.app_label, opts.module_name
+            app_label, model_name = opts.app_label, opts.model_name
             return ('%s:%s_%s_simple_history' %
                     (admin.site.name, app_label, model_name),
                     [getattr(self, opts.pk.attname), self.history_id])
 
         def get_instance(self):
-            return model(**dict([(field.attname, getattr(self, field.attname))
-                                for field in fields.values()]))
+            return model(**{
+                field.attname: getattr(self, field.attname)
+                for field in fields.values()
+            })
 
         return {
             'history_id': models.AutoField(primary_key=True),
@@ -356,81 +369,6 @@ class HistoricalRecords(object):
                 return None
             except AttributeError:
                 return None
-
-
-class CustomForeignKeyField(models.ForeignKey):
-
-    def __init__(self, *args, **kwargs):
-        warnings.warn("CustomForeignKeyField is deprecated.",
-                      DeprecationWarning)
-        super(CustomForeignKeyField, self).__init__(*args, **kwargs)
-        self.db_constraint = False
-        self.generate_reverse_relation = False
-
-    def get_attname(self):
-        return self.name
-
-    def get_one_to_one_field(self, to_field, other):
-        # HACK This creates a new custom foreign key based on to_field,
-        # and calls itself with that, effectively making the calls
-        # recursive
-        temp_field = self.__class__(to_field.rel.to._meta.object_name)
-        for key, val in to_field.__dict__.items():
-            if (isinstance(key, six.string_types)
-                    and not key.startswith('_')):
-                setattr(temp_field, key, val)
-        field = self.__class__.get_field(
-            temp_field, other, to_field.rel.to)
-        return field
-
-    def get_field(self, other, cls):
-        # this hooks into contribute_to_class() and this is
-        # called specifically after the class_prepared signal
-        to_field = copy.copy(self.rel.to._meta.pk)
-        field = self
-        if isinstance(to_field, models.OneToOneField):
-            field = self.get_one_to_one_field(to_field, other)
-        elif isinstance(to_field, models.AutoField):
-            field.__class__ = convert_auto_field(to_field)
-        else:
-            field.__class__ = to_field.__class__
-            excluded_prefixes = ("_", "__")
-            excluded_attributes = (
-                "rel",
-                "creation_counter",
-                "validators",
-                "error_messages",
-                "attname",
-                "column",
-                "help_text",
-                "name",
-                "model",
-                "unique_for_year",
-                "unique_for_date",
-                "unique_for_month",
-                "db_tablespace",
-                "db_index",
-                "db_column",
-                "default",
-                "auto_created",
-                "null",
-                "blank",
-            )
-            for key, val in to_field.__dict__.items():
-                if (isinstance(key, six.string_types)
-                        and not key.startswith(excluded_prefixes)
-                        and key not in excluded_attributes):
-                    setattr(field, key, val)
-        return field
-
-    def do_related_class(self, other, cls):
-        field = self.get_field(other, cls)
-        transform_field(field)
-        field.rel = None
-
-    def contribute_to_class(self, cls, name):
-        # HACK: remove annoying descriptor (don't super())
-        RelatedField.contribute_to_class(self, cls, name)
 
 
 def transform_field(field):