+++ simple_history/tests/tests/test_models.py
@@ -2,26 +2,36 @@ from __future__ import unicode_literals
 
 from datetime import datetime, timedelta
 import unittest
+import warnings
 
 import django
 from django.db import models
-from django.db.models.loading import get_model
 from django.db.models.fields.proxy import OrderWrt
 from django.test import TestCase
 from django.core.files.base import ContentFile
 
+from simple_history import exceptions, register
 from simple_history.models import HistoricalRecords, convert_auto_field
-from simple_history import register
 from ..models import (
     AdminProfile, Bookcase, MultiOneToOne, Poll, Choice, Voter, Restaurant,
     Person, FileModel, Document, Book, HistoricalPoll, Library, State,
     AbstractBase, ConcreteAttr, ConcreteUtil, SelfFK, Temperature, WaterLevel,
     ExternalModel1, ExternalModel3, UnicodeVerboseName, HistoricalChoice,
     HistoricalState, HistoricalCustomFKError, Series, SeriesWork, PollInfo,
-    UserAccessorDefault, UserAccessorOverride, Employee
+    UserAccessorDefault, UserAccessorOverride, Employee, Country, Province,
+    City, Contact, ContactRegister,
+    TrackedAbstractBaseA, TrackedAbstractBaseB,
+    TrackedWithAbstractBase, TrackedWithConcreteBase,
+    InheritTracking1, InheritTracking2, InheritTracking3, InheritTracking4,
 )
 from ..external.models import ExternalModel2, ExternalModel4
 
+try:
+    from django.apps import apps
+except ImportError:  # Django < 1.7
+    from django.db.models import get_model
+else:
+    get_model = apps.get_model
 try:
     from unittest import skipUnless
 except ImportError:
@@ -257,6 +267,22 @@ class HistoricalRecordsTest(TestCase):
         self.assertEqual([m.fk_id for m in model.history.all()],
                          [other.id, model.id, None])
 
+    def test_to_field_foreign_key_save(self):
+        country = Country.objects.create(code='US')
+        country2 = Country.objects.create(code='CA')
+        province = Province.objects.create(country=country)
+        province.country = country2
+        province.save()
+        self.assertEqual([c.country_id for c in province.history.all()],
+                         [country2.code, country.code])
+
+    def test_db_column_foreign_key_save(self):
+        country = Country.objects.create(code='US')
+        city = City.objects.create(country=country)
+        country_field = City._meta.get_field('country')
+        self.assertIn(getattr(country_field, 'db_column'),
+                      str(city.history.all().query))
+
     def test_raw_save(self):
         document = Document()
         document.save_base(raw=True)
@@ -309,12 +335,8 @@ class RegisterTest(TestCase):
         self.assertEqual(len(user.histories.all()), 1)
 
     def test_reregister(self):
-        register(Restaurant, manager_name='again')
-        register(User, manager_name='again')
-        self.assertTrue(hasattr(Restaurant, 'updates'))
-        self.assertFalse(hasattr(Restaurant, 'again'))
-        self.assertTrue(hasattr(User, 'histories'))
-        self.assertFalse(hasattr(User, 'again'))
+        with self.assertRaises(exceptions.MultipleRegistrationsError):
+            register(Restaurant, manager_name='again')
 
     def test_register_custome_records(self):
         self.assertEqual(len(Voter.history.all()), 0)
@@ -324,7 +346,8 @@ class RegisterTest(TestCase):
         voter = Voter.objects.create(choice=choice, user=user)
         self.assertEqual(len(voter.history.all()), 1)
         expected = 'Voter object changed by None as of '
-        self.assertEqual(expected, str(voter.history.all()[0])[:len(expected)])
+        self.assertEqual(expected,
+                         str(voter.history.all()[0])[:len(expected)])
 
 
 class CreateHistoryModelTests(unittest.TestCase):
@@ -490,7 +513,7 @@ class HistoryManagerTest(TestCase):
             self.assertRaises(TypeError, HistoricalRecords, bases=bases)
 
     def test_import_related(self):
-        field_object = HistoricalChoice._meta.get_field_by_name('poll')[0]
+        field_object = HistoricalChoice._meta.get_field('poll')
         try:
             related_model = field_object.rel.related_model
         except AttributeError:  # Django<1.8
@@ -498,7 +521,7 @@ class HistoryManagerTest(TestCase):
         self.assertEqual(related_model, HistoricalChoice)
 
     def test_string_related(self):
-        field_object = HistoricalState._meta.get_field_by_name('library')[0]
+        field_object = HistoricalState._meta.get_field('library')
         try:
             related_model = field_object.rel.related_model
         except AttributeError:  # Django<1.8
@@ -525,9 +548,13 @@ class TestConvertAutoField(TestCase):
 
         Default Django ORM uses an integer-based auto field.
         """
-        with self.settings(DATABASES={'default': {
-                'ENGINE': 'django.db.backends.postgresql_psycopg2'}}):
-            assert convert_auto_field(self.field) == models.IntegerField
+        with warnings.catch_warnings():
+            warnings.filterwarnings('ignore', category=UserWarning,
+                                    message='Overriding setting DATABASES ' +
+                                            'can lead to unexpected behavior.')
+            with self.settings(DATABASES={'default': {
+                    'ENGINE': 'django.db.backends.postgresql_psycopg2'}}):
+                assert convert_auto_field(self.field) == models.IntegerField
 
     def test_non_relational(self):
         """Non-relational test
@@ -535,9 +562,14 @@ class TestConvertAutoField(TestCase):
         MongoDB uses a string-based auto field. We need to make sure
         the converted field type is string.
         """
-        with self.settings(DATABASES={'default': {
-                'ENGINE': 'django_mongodb_engine'}}):
-            assert convert_auto_field(self.field) == models.TextField
+
+        with warnings.catch_warnings():
+            warnings.filterwarnings('ignore', category=UserWarning,
+                                    message='Overriding setting DATABASES ' +
+                                            'can lead to unexpected behavior.')
+            with self.settings(DATABASES={'default': {
+                    'ENGINE': 'django_mongodb_engine'}}):
+                assert convert_auto_field(self.field) == models.TextField
 
 
 class TestOrderWrtField(TestCase):
@@ -577,7 +609,7 @@ class TestOrderWrtField(TestCase):
             self.w_horse.pk,
             self.w_nephew.pk,
             self.w_battle.pk]
-        self.assertEqual(order, expected)
+        self.assertSequenceEqual(order, expected)
         self.assertEqual(0, self.w_lion._order)
         self.assertEqual(1, self.w_caspian._order)
         self.assertEqual(2, self.w_voyage._order)
@@ -587,11 +619,11 @@ class TestOrderWrtField(TestCase):
         self.assertEqual(6, self.w_battle._order)
 
     def test_order_field_in_historical_model(self):
-        work_order_field = self.w_lion._meta.get_field_by_name('_order')[0]
+        work_order_field = self.w_lion._meta.get_field('_order')
         self.assertEqual(type(work_order_field), OrderWrt)
 
         history = self.w_lion.history.all()[0]
-        history_order_field = history._meta.get_field_by_name('_order')[0]
+        history_order_field = history._meta.get_field('_order')
         self.assertEqual(type(history_order_field), models.IntegerField)
 
     def test_history_object_has_order(self):
@@ -615,7 +647,8 @@ class TestOrderWrtField(TestCase):
             self.w_chair.pk,
             self.w_battle.pk]
         self.series.set_serieswork_order(chronological)
-        self.assertEqual(self.series.get_serieswork_order(), chronological)
+        self.assertSequenceEqual(self.series.get_serieswork_order(),
+                                 chronological)
 
         # This uses an update, not a save, so no new history is created
         w_caspian = SeriesWork.objects.get(id=self.w_caspian.id)
@@ -704,7 +737,8 @@ class TestUserAccessor(unittest.TestCase):
         assert not hasattr(User, 'historicaluseraccessordefault_set')
 
     def test_accessor_override(self):
-        register(UserAccessorOverride, user_related_name='my_history_model_accessor')
+        register(UserAccessorOverride,
+                 user_related_name='my_history_model_accessor')
         assert hasattr(User, 'my_history_model_accessor')
 
 
@@ -729,3 +763,68 @@ class TestMissingOneToOne(TestCase):
         self.assertEqual(original.manager_id, 1)
         with self.assertRaises(Employee.DoesNotExist):
             original.manager
+
+
+class CustomTableNameTest1(TestCase):
+
+    @staticmethod
+    def get_table_name(manager):
+        return manager.model._meta.db_table
+
+    def test_custom_table_name(self):
+        self.assertEqual(
+            self.get_table_name(Contact.history),
+            'contacts_history',
+        )
+
+    def test_custom_table_name_from_register(self):
+        self.assertEqual(
+            self.get_table_name(ContactRegister.history),
+            'contacts_register_history',
+        )
+
+
+class TestTrackingInheritance(TestCase):
+
+    def test_tracked_abstract_base(self):
+        self.assertEqual(
+            [f.attname for f in TrackedWithAbstractBase.history.model._meta.fields],
+            ['id', 'history_id', 'history_date', 'history_user_id', 'history_type'],
+        )
+
+    def test_tracked_concrete_base(self):
+        self.assertEqual(
+            [f.attname for f in TrackedWithConcreteBase.history.model._meta.fields],
+            ['id', 'trackedconcretebase_ptr_id', 'history_id', 'history_date', 'history_user_id', 'history_type'],
+        )
+
+    def test_multiple_tracked_bases(self):
+        with self.assertRaises(exceptions.MultipleRegistrationsError):
+            class TrackedWithMultipleAbstractBases(TrackedAbstractBaseA, TrackedAbstractBaseB):
+                pass
+
+    def test_tracked_abstract_and_untracked_concrete_base(self):
+        self.assertEqual(
+            [f.attname for f in InheritTracking1.history.model._meta.fields],
+            ['id', 'untrackedconcretebase_ptr_id', 'history_id', 'history_date', 'history_user_id', 'history_type'],
+        )
+
+    def test_indirect_tracked_abstract_base(self):
+        self.assertEqual(
+            [f.attname for f in InheritTracking2.history.model._meta.fields],
+            [
+                'id', 'baseinherittracking2_ptr_id',
+                'history_id', 'history_date', 'history_user_id', 'history_type'],
+        )
+
+    def test_indirect_tracked_concrete_base(self):
+        self.assertEqual(
+            [f.attname for f in InheritTracking3.history.model._meta.fields],
+            [
+                'id', 'baseinherittracking3_ptr_id',
+                'history_id', 'history_date', 'history_user_id', 'history_type'],
+        )
+
+    def test_registering_with_tracked_abstract_base(self):
+        with self.assertRaises(exceptions.MultipleRegistrationsError):
+            register(InheritTracking4)